Add error handling example
This PR adds a recommended way to send errors back to the client-side.

The previous approach was brittle because if the following:
If required too much boilerplate code, meaning that all conditions
that could cause an error had to manipulate the context variable.

Also, we could have cases where the same error was being thrown with
different messages.

I have also added a code error, so that frontend can use that
as an ENUM to take different actions depending on the error type.
This is more powerful than just using HTTP status.

At last, another advantage of this is that it is dead simple to change
the error messages format if we need in the future. Everything is
contained inside the error function and the errorHandler middleware.